requirements.txt
langchain==0.1.1
python-dotenv==1.0.0
tiktoken==0.5.2
langchain-community==0.0.19
langchain-openai
faiss-cpu
rank_bm25